requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiomysql
alembic
pydantic[email]
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
email-validator
python-multipart
APScheduler
fastapi-mail
httpx
